+++ src/HasResult.js
@@ -1,21 +1,23 @@
 import * as PropTypes from 'prop-types'
 import invariant from 'fbjs/lib/invariant'
+import { resultStoreContextPropType, resultStoreContextPropName } from './common'
 
 export const createHasResult = (contextPropName, rootDisplayName) => {
   const HasResult = (props, context) => {
     invariant(
-      context[contextPropName],
+      context[contextPropName] && context[contextPropName][resultStoreContextPropName],
       `<${HasResult.displayName}> must be a child (direct or indirect) of <${rootDisplayName}>.`,
     )
 
-    return context[contextPropName].hasResult ? props.children(context[contextPropName].result) : null
+    return (
+      (context[contextPropName][resultStoreContextPropName].hasResult &&
+        props.children({ result: context[contextPropName][resultStoreContextPropName].result })) ||
+      null
+    )
   }
 
   HasResult.contextTypes = {
-    [contextPropName]: PropTypes.shape({
-      hasResult: PropTypes.bool,
-      result: PropTypes.any,
-    }),
+    [contextPropName]: resultStoreContextPropType,
   }
 
   HasResult.propTypes = {