+++ src/__tests__/AsyncCall.test.js
@@ -3,8 +3,7 @@ import * as PropTypes from 'prop-types'
 import { shallow, mount } from 'enzyme'
 
 import createAsyncCallComponent from '../'
-
-const flushPromises = () => new Promise(resolve => setImmediate(resolve))
+import { flushPromises } from './common'
 
 describe('AsyncCall', () => {
   it('should throw an error if a function is not passed to createAsyncCallComponent', () => {
@@ -106,7 +105,6 @@ describe('AsyncCall', () => {
       running: true,
       resolved: false,
       rejected: false,
-      hasResult: false,
       execute: container.instance().execute,
     })
   })
@@ -116,14 +114,6 @@ describe('AsyncCall', () => {
     const children = jest.fn(() => null)
     const container = shallow(<AsyncCall params={{}}>{children}</AsyncCall>)
 
-    expect(children).toHaveBeenLastCalledWith({
-      running: true,
-      resolved: false,
-      rejected: false,
-      hasResult: false,
-      execute: container.instance().execute,
-    })
-
     await flushPromises()
 
     expect(children).toHaveBeenLastCalledWith({
@@ -131,7 +121,6 @@ describe('AsyncCall', () => {
       resolved: true,
       result: 42,
       rejected: false,
-      hasResult: true,
       execute: container.instance().execute,
     })
 
@@ -143,14 +132,6 @@ describe('AsyncCall', () => {
     const children = jest.fn(() => null)
     const container = shallow(<AsyncCall params={{}}>{children}</AsyncCall>)
 
-    expect(children).toHaveBeenLastCalledWith({
-      running: true,
-      rejected: false,
-      resolved: false,
-      hasResult: false,
-      execute: container.instance().execute,
-    })
-
     await flushPromises()
 
     expect(children).toHaveBeenLastCalledWith({
@@ -158,7 +139,6 @@ describe('AsyncCall', () => {
       rejected: true,
       resolved: false,
       rejectReason: 'rejected',
-      hasResult: false,
       execute: container.instance().execute,
     })
 
@@ -184,26 +164,6 @@ describe('AsyncCall', () => {
     expect(container.childAt(1).text()).toBe('12345')
   })
 
-  it('should call children fn and pass { running: true, rejected: false, result: <previous result> } as an arguments to it if promise has been called the second time after resolving', async done => {
-    const AsyncCall = createAsyncCallComponent(value => Promise.resolve(value))
-    const children = jest.fn(() => null)
-    const container = shallow(<AsyncCall params="first">{children}</AsyncCall>)
-
-    await flushPromises()
-    container.setProps({ params: 'second' })
-
-    expect(children).toHaveBeenLastCalledWith({
-      running: true,
-      rejected: false,
-      resolved: false,
-      result: 'first',
-      hasResult: true,
-      execute: container.instance().execute,
-    })
-
-    done()
-  })
-
   it('should call children fn and pass { running: true, rejected: false, rejectedResult: undefined } as an arguments to it if promise has been called the second time after rejection', async done => {
     const AsyncCall = createAsyncCallComponent(() => Promise.reject('rejected'))
     const children = jest.fn(() => null)
@@ -217,57 +177,10 @@ describe('AsyncCall', () => {
       rejected: false,
       resolved: false,
       rejectReason: undefined,
-      hasResult: false,
       execute: container.instance().execute,
     })
 
     done()
-  })
-
-  describe('mergeResult', () => {
-    it('should not call mergeResult callback if promise has been called the first time', async done => {
-      const AsyncCall = createAsyncCallComponent(value => Promise.resolve())
-      const mergeResult = jest.fn(() => void 0)
-      const container = shallow(<AsyncCall params={0} mergeResult={mergeResult} />)
-      await flushPromises()
-
-      expect(mergeResult).not.toHaveBeenCalled()
-
-      done()
-    })
-
-    it('should call mergeResult callback if promise has been called the second time', async done => {
-      const AsyncCall = createAsyncCallComponent(value => Promise.resolve(value))
-      const mergeResult = jest.fn((prevResult, currentResult) => void 0)
-      const container = shallow(<AsyncCall params={0} mergeResult={mergeResult} />)
-
-      container.setProps({ params: 1 })
-
-      await flushPromises()
-
-      expect(mergeResult).toHaveBeenCalledTimes(1)
-      expect(mergeResult).toHaveBeenCalledWith(0, 1)
-
-      done()
-    })
-
-    it('should pass returning value of mergeResult callback to a children function', async done => {
-      const AsyncCall = createAsyncCallComponent(value => Promise.resolve(value))
-      const mergeResult = (prevResult, currentResult) => prevResult + currentResult
-      const children = jest.fn(result => null)
-      const container = shallow(
-        <AsyncCall params={10} mergeResult={mergeResult}>
-          {({ result }) => children(result)}
-        </AsyncCall>,
-      )
-
-      container.setProps({ params: 32 })
-
-      await flushPromises()
-      expect(children).toHaveBeenLastCalledWith(42)
-
-      done()
-    })
   })
 
   describe('context', () => {
@@ -300,7 +213,6 @@ describe('AsyncCall', () => {
         </AsyncCall>,
       )
       expect(rootContext).toBeDefined()
-      expect(rootContext).toHaveProperty('hasResult')
       expect(rootContext).toHaveProperty('running')
       expect(rootContext).toHaveProperty('rejected')
       expect(rootContext).toHaveProperty('resolved')
@@ -318,7 +230,6 @@ describe('AsyncCall', () => {
           <ContextChecker />
         </AsyncCall>,
       )
-      expect(rootContext.hasResult).toBe(false)
       expect(rootContext.running).toBe(true)
       expect(rootContext.rejected).toBe(false)
       expect(rootContext.resolved).toBe(false)
@@ -339,7 +250,6 @@ describe('AsyncCall', () => {
 
       await flushPromises()
 
-      expect(rootContext.hasResult).toBe(true)
       expect(rootContext.running).toBe(false)
       expect(rootContext.rejected).toBe(false)
       expect(rootContext.resolved).toBe(true)
@@ -361,7 +271,6 @@ describe('AsyncCall', () => {
 
       await flushPromises()
 
-      expect(rootContext.hasResult).toBe(false)
       expect(rootContext.running).toBe(false)
       expect(rootContext.rejected).toBe(true)
       expect(rootContext.resolved).toBe(false)
@@ -384,12 +293,11 @@ describe('AsyncCall', () => {
       await flushPromises()
       container.setProps({ params: 'second' })
 
-      expect(rootContext.hasResult).toBe(true)
       expect(rootContext.running).toBe(true)
       expect(rootContext.rejected).toBe(false)
       expect(rootContext.resolved).toBe(false)
       expect(rootContext.execute).toBe(container.instance().execute)
-      expect(rootContext.result).toBe('first')
+      expect(rootContext).not.toHaveProperty('result')
       expect(rootContext).not.toHaveProperty('rejectReason')
 
       done()
@@ -407,7 +315,6 @@ describe('AsyncCall', () => {
       await flushPromises()
       container.setProps({ params: { a: 2 } })
 
-      expect(rootContext.hasResult).toBe(false)
       expect(rootContext.running).toBe(true)
       expect(rootContext.rejected).toBe(false)
       expect(rootContext.resolved).toBe(false)