+++ src/__tests__/HasResult.test.js
@@ -2,37 +2,49 @@ import * as React from 'react'
 import { shallow, mount } from 'enzyme'
 
 import createAsyncCallComponent from '../index'
+import { flushPromises } from './common'
 
-const flushPromises = () => new Promise(resolve => setImmediate(resolve))
-
-describe('Result', () => {
+describe('HasResult', () => {
   it('should be exposed as static prop from AsyncCall', () => {
     const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
-    expect(AsyncCall.HasResult).toBeDefined()
+    expect(AsyncCall.ResultStore.HasResult).toBeDefined()
   })
 
   it('should expose default display names', () => {
     const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
-    expect(AsyncCall.HasResult.displayName).toBe('AsyncCall.HasResult')
+    expect(AsyncCall.ResultStore.HasResult.displayName).toBe('AsyncCall.ResultStore.HasResult')
   })
 
   it('should throw an error if Result component rendered alone', () => {
     const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
-    expect(() => shallow(<AsyncCall.HasResult>{() => {}}</AsyncCall.HasResult>)).toThrow(
-      '<AsyncCall.HasResult> must be a child (direct or indirect) of <AsyncCall>.',
+    expect(() => shallow(<AsyncCall.ResultStore.HasResult>{() => null}</AsyncCall.ResultStore.HasResult>)).toThrow(
+      '<AsyncCall.ResultStore.HasResult> must be a child (direct or indirect) of <AsyncCall.ResultStore>.',
     )
   })
 
-  // The test below is disabled for now because jest do not catch React errors properly
+  // The tests below is disabled for now because jest do not catch React errors properly
   // See the following issues for further details:
   // https://github.com/facebook/react/issues/11098
   // https://github.com/airbnb/enzyme/issues/1280
+  xit('should throw an error if Result component rendered as a direct child of <AsyncCall>', () => {
+    const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
+    expect(() =>
+      mount(
+        <AsyncCall params={1}>
+          <AsyncCall.ResultStore.HasResult>{() => null}</AsyncCall.ResultStore.HasResult>
+        </AsyncCall>,
+      ),
+    ).toThrow('<AsyncCall.ResultStore.HasResult> must be a child (direct or indirect) of <AsyncCall.ResultStore>.')
+  })
+
   xit('should throw an error if children is not passed', () => {
     const AsyncCall = createAsyncCallComponent(value => Promise.resolve(value))
     expect(() =>
       mount(
         <AsyncCall params="first">
-          <AsyncCall.HasResult />
+          <AsyncCall.ResultStore>
+            <AsyncCall.ResultStore.HasResult>{() => null}</AsyncCall.ResultStore.HasResult>
+          </AsyncCall.ResultStore>
         </AsyncCall>,
       ),
     ).toThrow()
@@ -43,12 +55,18 @@ describe('Result', () => {
     const children = jest.fn(() => null)
     const container = mount(
       <AsyncCall params={{}}>
-        <AsyncCall.HasResult>{children}</AsyncCall.HasResult>
+        <AsyncCall.ResultStore>
+          <AsyncCall.ResultStore.HasResult>{children}</AsyncCall.ResultStore.HasResult>
+        </AsyncCall.ResultStore>
       </AsyncCall>,
     )
 
     expect(container.children().exists()).toBe(true)
-    const resultContainer = container.childAt(0)
+    const resultStoreContainer = container.childAt(0)
+    expect(resultStoreContainer).toBeDefined()
+
+    expect(resultStoreContainer.children().exists()).toBe(true)
+    const resultContainer = resultStoreContainer.childAt(0)
     expect(resultContainer).toBeDefined()
     expect(resultContainer).toHaveEmptyRender()
 
@@ -60,16 +78,23 @@ describe('Result', () => {
     const children = jest.fn(() => null)
     const container = mount(
       <AsyncCall params={{}}>
-        <AsyncCall.HasResult>{children}</AsyncCall.HasResult>
+        <AsyncCall.ResultStore>
+          <AsyncCall.ResultStore.HasResult>{children}</AsyncCall.ResultStore.HasResult>
+        </AsyncCall.ResultStore>
       </AsyncCall>,
     )
 
     await flushPromises()
 
     expect(container.children().exists()).toBe(true)
-    const resultContainer = container.childAt(0)
+    const resultStoreContainer = container.childAt(0)
+    expect(resultStoreContainer).toBeDefined()
+
+    expect(resultStoreContainer.children().exists()).toBe(true)
+    const resultContainer = resultStoreContainer.childAt(0)
     expect(resultContainer).toBeDefined()
     expect(resultContainer).toHaveEmptyRender()
+
     expect(children).not.toHaveBeenCalled()
 
     done()
@@ -80,14 +105,16 @@ describe('Result', () => {
     const children = jest.fn(value => <div>result</div>)
     const container = mount(
       <AsyncCall params={{}}>
-        <AsyncCall.HasResult>{children}</AsyncCall.HasResult>
+        <AsyncCall.ResultStore>
+          <AsyncCall.ResultStore.HasResult>{children}</AsyncCall.ResultStore.HasResult>
+        </AsyncCall.ResultStore>
       </AsyncCall>,
     )
 
     await flushPromises()
     container.update()
 
-    expect(children).toHaveBeenCalledWith(42)
+    expect(children).toHaveBeenCalledWith({ result: 42 })
 
     expect(container.children().exists()).toBe(true)
     const resultContainer = container.childAt(0)