+++ src/__tests__/Resolved.test.js
@@ -2,8 +2,7 @@ import * as React from 'react'
 import { shallow, mount } from 'enzyme'
 
 import createAsyncCallComponent from '../index'
-
-const flushPromises = () => new Promise(resolve => setImmediate(resolve))
+import { flushPromises } from './common'
 
 describe('Resolved', () => {
   it('should be exposed as static prop from AsyncCall', () => {
@@ -23,6 +22,18 @@ describe('Resolved', () => {
     )
   })
 
+  it("should not call Resolved's children function if promise has not been resolved yet", () => {
+    const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
+    const children = jest.fn(() => void 0)
+    const container = mount(
+      <AsyncCall params={{}}>
+        <AsyncCall.Resolved>{children}</AsyncCall.Resolved>
+      </AsyncCall>,
+    )
+
+    expect(children).not.toHaveBeenCalled()
+  })
+
   it("should not render Resolved's children if promise has not been resolved yet", () => {
     const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
     const container = mount(
@@ -35,6 +46,20 @@ describe('Resolved', () => {
     const resolvedContainer = container.childAt(0)
     expect(resolvedContainer).toBeDefined()
     expect(resolvedContainer).toHaveEmptyRender()
+  })
+
+  it("should not call Resolved's children function if promise has been rejected", async done => {
+    const AsyncCall = createAsyncCallComponent(() => Promise.reject('error'))
+    const children = jest.fn(() => void 0)
+    mount(
+      <AsyncCall params={{}}>
+        <AsyncCall.Resolved>{children}</AsyncCall.Resolved>
+      </AsyncCall>,
+    )
+
+    await flushPromises()
+    expect(children).not.toHaveBeenCalled()
+    done()
   })
 
   it("should not render Resolved's children if promise has been rejected", async done => {
@@ -53,6 +78,71 @@ describe('Resolved', () => {
 
     done()
   })
+
+  it("should call Resolved's children function if promise has been resolved", async done => {
+    const AsyncCall = createAsyncCallComponent(() => Promise.resolve(42))
+    const children = jest.fn(() => <div>ABCDEF</div>)
+    const container = mount(
+      <AsyncCall params={{}}>
+        <AsyncCall.Resolved>{children}</AsyncCall.Resolved>
+      </AsyncCall>,
+    )
+
+    await flushPromises()
+    container.update()
+
+    expect(children).toHaveBeenCalledWith({ result: 42 })
+    expect(container.children().exists()).toBe(true)
+    const resolvedContainer = container.childAt(0)
+    expect(resolvedContainer).toBeDefined()
+    expect(resolvedContainer).not.toHaveEmptyRender()
+    expect(resolvedContainer.text()).toBe('ABCDEF')
+
+    done()
+  })
+
+  it("should render Resolved's empty children if promise has been resolved", async done => {
+    const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
+    const container = mount(
+      <AsyncCall params={{}}>
+        <AsyncCall.Resolved />
+      </AsyncCall>,
+    )
+
+    await flushPromises()
+    container.update()
+
+    expect(container.children().exists()).toBe(true)
+    const resolvedContainer = container.childAt(0)
+
+    expect(resolvedContainer).toBeDefined()
+    expect(resolvedContainer).toHaveEmptyRender()
+
+    done()
+  })
+
+  const values = [false, null, undefined]
+  values.forEach(value =>
+    it(`should render Resolved's children function that returns ${value} if promise has been resolved`, async done => {
+      const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
+      const container = mount(
+        <AsyncCall params={{}}>
+          <AsyncCall.Resolved>{() => value}</AsyncCall.Resolved>
+        </AsyncCall>,
+      )
+
+      await flushPromises()
+      container.update()
+
+      expect(container.children().exists()).toBe(true)
+      const resolvedContainer = container.childAt(0)
+
+      expect(resolvedContainer).toBeDefined()
+      expect(resolvedContainer).toHaveEmptyRender()
+
+      done()
+    }),
+  )
 
   it("should render Resolved's children if promise has been resolved", async done => {
     const AsyncCall = createAsyncCallComponent(() => Promise.resolve())
@@ -100,6 +190,23 @@ describe('Resolved', () => {
     done()
   })
 
+  it("should not call Resolved's children function if promise has not been resolved the second time", async done => {
+    const fn = jest.fn(value => Promise.resolve(value))
+    const children = jest.fn(result => null)
+    const AsyncCall = createAsyncCallComponent(fn)
+    const container = mount(
+      <AsyncCall params="first">
+        <AsyncCall.Resolved>{children}</AsyncCall.Resolved>
+      </AsyncCall>,
+    )
+
+    await flushPromises()
+    container.instance().execute()
+    expect(children).toHaveBeenCalledTimes(1)
+
+    done()
+  })
+
   it("should not render Resolved's children if promise has not been resolved the second time", async done => {
     const fn = jest.fn(value => Promise.resolve(value))
     const AsyncCall = createAsyncCallComponent(fn)
@@ -110,7 +217,6 @@ describe('Resolved', () => {
     )
 
     await flushPromises()
-
     container.instance().execute()
 
     expect(fn).toHaveBeenCalledTimes(2)